Refactor for centralized configuration
